UISAUTHCOM-15: move assign users flow parts to separate hook
