+++ lib/hooks/index.js
@@ -27,3 +27,4 @@ export * from './useUsers';
 export * from './useUserCapabilities';
 export * from './useUserCapabilitiesSets';
 export * from './useCreateAuthUserKeycloak';
+export * from './useAssignUsersFlow';
